Класс Field вынесен в отдельные файлы
Обновлены проекты code::blocks и cmake
Функции общего назначения вынесены в Common
